Remove System.Data.Common package dependency
System.Data.Common is part of .NET and the package dependency is not needed. The package dependency also generates false positive security warnings in some scanning tools, as it references packages with known security issues. It is a false positive because when the application is deployed, the current version from .NET is used, not version 4.3.0.