Make all params named params
Update the `BatchBuilder` to add all positional parameters as named parameters. This is the same behavior as `CommandBuilder`.

Under the hood what is going on is that if there is a mix of positional and named parameters, then npgsql does not edit the query to normalise all parameters. This results in postgres throwing an error (cant mix positional and named parameters).  If there were only positional parameters (old behhaviour), the npgsql will rewrite the query to convert all positional params to named params. This is done for maximum compatibility.  The easiest way to make things consistently work is to make all positional params as named params from the start, which is what `CommandBuilder`/`CommandBuilderBase` does.

This _may_ be a breaking change - but only if people were already using named parameters with the batch builder and only if they were naming their params `p0`, `p1`, etc.